Log killler info
Adds the killer information to the Hardcore Death Log, and displays it in the Log frame. Note that the addon will error attempting to display the Log frame until you delete WTF\Account\<AccountName>\SavedVariables\HardcoreDeath.lua, or hand-edit this file to remove log entries that do not contain killer information.